requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
psycopg2-binary==2.9.11
rapidfuzz==3.5.2
feedparser==6.0.11
requests==2.31.0
dotenv==0.9.9
gunicorn==23.0.0
numpy==1.26.4